pong v0.12

+++ back/src/modules/game/game.event.ts
@@ -9,6 +9,9 @@ import {
 } from '@nestjs/websockets';
 import { GameService } from './game.service';
 import { Game } from './game.service';
+import { UnauthorizedException } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import { JwtService } from '@nestjs/jwt';
 
 interface clientUpdate {
   posRacket: number;
@@ -27,46 +30,85 @@ export class GameEvents {
   @WebSocketServer()
   server: Server;
 
-  constructor(private gameService: GameService) {
+  constructor(
+    private gameService: GameService,
+    private configService: ConfigService,
+    private jwtService: JwtService,
+  ) {
     setInterval(() => {
       const games: Map<string, Game> = this.gameService.getGames();
       games.forEach((game) => {
         const update = this.gameService.updateGame(game);
-        this.server.to(game.player1Id).emit('gameUpdate', update);
+        this.server.to(game.socketPlayer1Id).emit('gameUpdate', update);
         update.isPlayerRight = true;
-        this.server.to(game.player2Id).emit('gameUpdate', update);
+        this.server.to(game.socketPlayer2Id).emit('gameUpdate', update);
       });
     }, 1000 / 60);
   }
 
   //conexion
-  handleConnection(client: Socket) {
+  async handleConnection(client: Socket) {
+    //const token = client.handshake.headers.cookies['jwt'];
+    let cookieArray = client.handshake.headers.cookie.split(';');
+    let jwtToken = '';
+    cookieArray.forEach((cookie) => {
+      let cookieParts = cookie.split('=');
+      if (cookieParts[0] === 'jwt') {
+        jwtToken = cookieParts[1];
+      }
+    });
+
+    if (!jwtToken) {
+      throw new UnauthorizedException("You're not logged in", 'No token found');
+    }
+
+    try {
+      const jwtSecret = this.configService.get<string>('JWT_SECRET');
+      const payload = await this.jwtService.verifyAsync(jwtToken, {
+        secret: jwtSecret,
+      });
+      // Maintenant, vous avez validé le jeton et extrait le payload.
+      // Vous pouvez utiliser le payload pour obtenir des informations sur l'utilisateur.
+
+      // Par exemple, si vous avez l'ID de l'utilisateur dans le payload :
+      client.data.userId = payload.login;
+      console.log(client.data);
+    } catch (e) {
+      throw new UnauthorizedException('Authorization error', e.message);
+    }
     console.log('client connected: ' + client.id);
-    //this.gameService.checkAlreadyInGame(client.id);
+    if (this.gameService.checkAlreadyInGame(client.id))
+      this.server.to(client.id).emit('userGameStatus', 'alreadyInGame');
+    return 'connected';
   }
 
-  //desconexion
+  //desconexio
   handleDisconnect(client: Socket) {
     console.log('client disconnected: ' + client.id);
     this.gameService.removeFromQueue(client.id);
   }
 
-  @SubscribeMessage('searchOpponent')
+  @SubscribeMessage('userGameStatus')
   handleSearchOpponent(
     @MessageBody() message: string,
     @ConnectedSocket() client: Socket,
   ) {
-    console.log('searchOpponent: ' + message);
+    if (!client.data.userId) {
+      return 'error';
+    }
     if (message === 'cancel') {
       this.gameService.removeFromQueue(client.id);
       return;
-    } else {
-      const opponent = this.gameService.addToQueue(client.id);
+    } else if (message === 'search') {
+      const opponent = this.gameService.addToQueue(
+        client.id,
+        client.data.userId,
+      );
       console.log('searching opponent for: ' + client.id);
       if (opponent) {
         console.log('opponent found: ' + opponent);
-        this.server.to(client.id).emit('searchOpponent', opponent);
-        this.server.to(opponent).emit('searchOpponent', client.id);
+        this.server.to(client.id).emit('userGameStatus', 'found');
+        this.server.to(opponent).emit('userGameStatus', 'found');
       }
     }
   }
